pyproject.toml
[project]
name = "neura-craft-backend"
version = "0.1.0"
description = "FastAPI backend for Neura Craft"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Neura Craft Backend Team" }
]
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "pyodbc>=5.1.0",
    "alembic>=1.13.2",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "azure-identity>=1.17.1",
    "azure-storage-blob>=12.19.0",
    "httpx>=0.27.0",
    "aiofiles>=23.2.1",
    "email-validator>=2.0.0",
    "python-multipart>=0.0.6"
]

[project.optional-dependencies]
test = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.0"
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-ra -q"
testpaths = ["tests"]


